[METRIC] all metric move finished
