+++ include/xgboost/sync.h
@@ -1,0 +1,13 @@
+/*!
+ * Copyright 2014 by Contributors
+ * \file sync.h
+ * \brief the synchronization module of rabit
+ *        redirects to rabit header
+ * \author Tianqi Chen
+ */
+#ifndef XGBOOST_SYNC_H_
+#define XGBOOST_SYNC_H_
+
+#include <rabit.h>
+
+#endif  // XGBOOST_SYNC_H_
