--- old_src/sync/sync.h
@@ -1,13 +1,0 @@
-/*!
- * Copyright 2014 by Contributors
- * \file sync.h
- * \brief the synchronization module of rabit
- *        redirects to subtree rabit header
- * \author Tianqi Chen
- */
-#ifndef XGBOOST_SYNC_SYNC_H_
-#define XGBOOST_SYNC_SYNC_H_
-
-#include "../../subtree/rabit/include/rabit.h"
-#include "../../subtree/rabit/include/rabit/timer.h"
-#endif  // XGBOOST_SYNC_SYNC_H_
