+++ src/metric/elementwise_metric.cc
@@ -1,0 +1,127 @@
+/*!
+ * Copyright 2015 by Contributors
+ * \file elementwise_metric.cc
+ * \brief evaluation metrics for elementwise binary or regression.
+ * \author Kailong Chen, Tianqi Chen
+ */
+#include <xgboost/metric.h>
+#include <xgboost/sync.h>
+#include <cmath>
+#include "../common/math.h"
+
+namespace xgboost {
+namespace metric {
+/*!
+ * \brief base class of element-wise evaluation
+ * \tparam Derived the name of subclass
+ */
+template<typename Derived>
+struct EvalEWiseBase : public Metric {
+  float Eval(const std::vector<float>& preds,
+             const MetaInfo& info,
+             bool distributed) const override {
+    CHECK_NE(info.labels.size(), 0) << "label set cannot be empty";
+    CHECK_EQ(preds.size(), info.labels.size())
+        << "label and prediction size not match, "
+        << "hint: use merror or mlogloss for multi-class classification";
+    const bst_omp_uint ndata = static_cast<bst_omp_uint>(info.labels.size());
+    float sum = 0.0, wsum = 0.0;
+    #pragma omp parallel for reduction(+: sum, wsum) schedule(static)
+    for (bst_omp_uint i = 0; i < ndata; ++i) {
+      const float wt = info.GetWeight(i);
+      sum += Derived::EvalRow(info.labels[i], preds[i]) * wt;
+      wsum += wt;
+    }
+    float dat[2]; dat[0] = sum, dat[1] = wsum;
+    if (distributed) {
+      rabit::Allreduce<rabit::op::Sum>(dat, 2);
+    }
+    return Derived::GetFinal(dat[0], dat[1]);
+  }
+  /*!
+   * \brief to be implemented by subclass,
+   *   get evaluation result from one row
+   * \param label label of current instance
+   * \param pred prediction value of current instance
+   */
+  inline static float EvalRow(float label, float pred);
+  /*!
+   * \brief to be overridden by subclass, final transformation
+   * \param esum the sum statistics returned by EvalRow
+   * \param wsum sum of weight
+   */
+  inline static float GetFinal(float esum, float wsum) {
+    return esum / wsum;
+  }
+};
+
+struct EvalRMSE : public EvalEWiseBase<EvalRMSE> {
+  const char *Name() const override {
+    return "rmse";
+  }
+  inline static float EvalRow(float label, float pred) {
+    float diff = label - pred;
+    return diff * diff;
+  }
+  inline static float GetFinal(float esum, float wsum) {
+    return std::sqrt(esum / wsum);
+  }
+};
+
+struct EvalLogLoss : public EvalEWiseBase<EvalLogLoss> {
+  const char *Name() const override {
+    return "logloss";
+  }
+  inline static float EvalRow(float y, float py) {
+    const float eps = 1e-16f;
+    const float pneg = 1.0f - py;
+    if (py < eps) {
+      return -y * std::log(eps) - (1.0f - y)  * std::log(1.0f - eps);
+    } else if (pneg < eps) {
+      return -y * std::log(1.0f - eps) - (1.0f - y)  * std::log(eps);
+    } else {
+      return -y * std::log(py) - (1.0f - y) * std::log(pneg);
+    }
+  }
+};
+
+struct EvalError : public EvalEWiseBase<EvalError> {
+  const char *Name() const override {
+    return "error";
+  }
+  inline static float EvalRow(float label, float pred) {
+    // assume label is in [0,1]
+    return pred > 0.5f ? 1.0f - label : label;
+  }
+};
+
+struct EvalPoissionNegLogLik : public EvalEWiseBase<EvalPoissionNegLogLik> {
+  const char *Name() const override {
+    return "poisson-nloglik";
+  }
+  inline static float EvalRow(float y, float py) {
+    const float eps = 1e-16f;
+    if (py < eps) py = eps;
+    return common::LogGamma(y + 1.0f) + py - std::log(py) * y;
+  }
+};
+
+XGBOOST_REGISTER_METRIC(RMSE, "rmse")
+.describe("Rooted mean square error.")
+.set_body([](const char* param) { return new EvalRMSE(); });
+
+XGBOOST_REGISTER_METRIC(LogLoss, "logloss")
+.describe("Negative loglikelihood for logistic regression.")
+.set_body([](const char* param) { return new EvalLogLoss(); });
+
+XGBOOST_REGISTER_METRIC(Error, "error")
+.describe("Binary classification error.")
+.set_body([](const char* param) { return new EvalError(); });
+
+XGBOOST_REGISTER_METRIC(PossionNegLoglik, "poisson-nloglik")
+.describe("Negative loglikelihood for poisson regression.")
+.set_body([](const char* param) { return new EvalPoissionNegLogLik(); });
+
+}  // namespace metric
+}  // namespace xgboost
+
