--- src/objective/objective.cc
@@ -1,17 +1,0 @@
-/*!
- * Copyright 2015 by Contributors
- * \file objective.cc
- * \brief global objective function definition.
- */
-#include <xgboost/objective.h>
-
-namespace dmlc {
-DMLC_REGISTRY_ENABLE(::xgboost::ObjFunctionReg);
-}  // namespace dmlc
-
-namespace xgboost {
-/*! \brief namespace of objective function */
-namespace obj {
-}  // namespace obj
-}  // namespace xgboost
-
